Update Set-CalendarProcessing.md
Based on our tests and confirmation from the engineering team (calendar team) after reviewing the code, this behavior is by design. In the principal/delegate scenario, if either the principal or the delegate has permission, the RBA will book. Please refer to the following VSO escalation for more details: 
https://o365exchange.visualstudio.com/DefaultCollection/Outlook%20Services/_workitems/edit/4459608 

We need to include this behavior in the -BookInPolicy property of Set-CalendarProcessing to avoid customer confusion and ensure their awareness.